+++ src/jesse_validator_draft2020_12.erl
@@ -979,7 +979,8 @@ check_required_values(Value, [PropertyName | Required], State) ->
 %% @doc maxProperties
 %% @private
 check_max_properties(Value, MaxProperties, State)
-  when is_integer(MaxProperties), MaxProperties >= 0 ->
+  when is_number(MaxProperties), MaxProperties >= 0,
+       trunc(MaxProperties) == MaxProperties ->
     case length(unwrap(Value)) =< MaxProperties of
       true  -> State;
       false -> handle_data_invalid(?too_many_properties, Value, State)
@@ -990,7 +991,8 @@ check_max_properties(_Value, _MaxProperties, State) ->
 %% @doc minProperties
 %% @private
 check_min_properties(Value, MinProperties, State)
-  when is_integer(MinProperties), MinProperties >= 0 ->
+  when is_number(MinProperties), MinProperties >= 0,
+       trunc(MinProperties) == MinProperties ->
     case length(unwrap(Value)) >= MinProperties of
       true  -> State;
       false -> handle_data_invalid(?too_few_properties, Value, State)